+++ packages/cli/templates/webcomponents/igc-ts/chip/default/index.ts
@@ -1,0 +1,19 @@
+import { TypeScriptFileUpdate } from "@igniteui/cli-core";
+import { IgniteUIForWebComponentsTemplate } from "../../../../../lib/templates/IgniteUIForWebComponentsTemplate";
+
+class IgcChipTemplate extends IgniteUIForWebComponentsTemplate {
+	constructor() {
+		super(__dirname);
+		this.components = ["Chip"];
+		this.controlGroup = "Data Entry & Display";
+		this.listInComponentTemplates = true;
+		this.id = "chip";
+		this.projectType = "igc-ts";
+		this.name = "Chip";
+		this.description = "Basic IgcChip";
+	}
+	protected addClassDeclaration(mainModule: TypeScriptFileUpdate, projPath: string, name: string, modulePath: string) {
+		// not applicable with custom module
+	}
+}
+module.exports = new IgcChipTemplate();