+++ packages/cli/templates/webcomponents/igc-ts/projects/side-nav/files/src/app/app.ts
@@ -9,7 +9,6 @@ import './app.module';
 export class App extends LitElement {
   static styles = css`
     router-outlet {
-      width: 100%;
       height: 100%;
       display: flex;
       text-align: center;
@@ -18,11 +17,15 @@ export class App extends LitElement {
       align-items: center;
       overflow: inherit;
     }
+
+    igc-nav-drawer {
+      float: left;
+    }
   `;
 
   render() {
     return html`
-      <igc-nav-drawer open=true>
+      <igc-nav-drawer open="true" position="relative">
         <igc-nav-drawer-header-item>Ignite UI CLI</igc-nav-drawer-header-item>
         ${routes.filter((element, index) => index < routes.length - 1).map(i => html`
           <igc-nav-drawer-item>