+++ spec/unit/PromptSession-spec.ts
@@ -693,11 +693,11 @@ describe("Unit - PromptSession", () => {
 		@use "igniteui-angular/theming" as *;
 		$primary: #731963 !default;
 		$secondary: #ce5712 !default;
-		$app-palette: igx-palette($primary, $secondary);
+		$app-palette: palette($primary, $secondary);
 		/* autoprefixer grid: on */
-		@include igx-core();
-		@include igx-typography($font-family: $material-typeface, $type-scale: $material-type-scale);
-		@include igx-theme($app-palette);`;
+		@include core();
+		@include typography($font-family: $material-typeface, $type-scale: $material-type-scale);
+		@include theme($app-palette);`;
 		const actualCall = (Util.processTemplates as jasmine.Spy).calls.argsFor(0)[2]["CustomTheme"].replace(/\s/g, "");
 		const expectedCall = CUSTOM_THEME.replace(/\s/g, "");
 		expect(actualCall).toEqual(expectedCall);