+++ CHANGELOG.md
@@ -1,3 +1,183 @@
+# [10.0.1](https://github.com/IgniteUI/igniteui-cli/compare/v10.0.0...v10.0.1) (2022-06-17)
+
+## igniteui/cli-core@10.0.1
+
+#### Bug Fixes
+
+* **package-manager:** skip setting up child install tasks for new projects ([de2a134](https://github.com/IgniteUI/igniteui-cli/commit/de2a1341e13ba20635a5ed94c5cc7e4c79c7ca46))
+
+## igniteui/angular-templates@14.0.1001
+
+* Note: Version bump only for package igniteui/cli-core
+
+## igniteui/angular-schematics@14.0.1001
+* Note: Version bump only for package igniteui/cli-core
+
+## igniteui-cli@10.0.1
+* Note: Version bump only for package igniteui/cli-core
+
+
+# [10.0.0](https://github.com/IgniteUI/igniteui-cli/compare/v9.2.0...v10.0.0)  (2022-06-09)
+
+🎉 This update includes:
+
+- `igniteui-angular` version [14.0.0](https://github.com/IgniteUI/igniteui-angular/releases/tag/14.0.0).
+
+## igniteui/cli-core@10.0.0
+
+* Note: Version bump only for package igniteui/cli-core
+
+## igniteui/angular-templates@14.0.1000
+
+* Update to igniteui-angular 14.0.0 by @dafo in https://github.com/IgniteUI/igniteui-cli/pull/1025
+
+## igniteui/angular-schematics@14.0.1000
+
+* Note: Version bump only for package igniteui/angular-schematics
+
+## igniteui-cli@10.0.0
+
+* Update to igniteui-angular 14.0.0 by @dafo in https://github.com/IgniteUI/igniteui-cli/pull/1025
+
+# [9.2.0](https://github.com/IgniteUI/igniteui-cli/compare/v9.1.4...v9.2.0)  (2022-05-26)
+
+🎉 This update includes:
+
+- `igniteui-angular` version [13.2.0](https://github.com/IgniteUI/igniteui-angular/releases/tag/13.2.0).
+
+## igniteui/cli-core@9.2.0
+
+* Note: Version bump only for package igniteui/cli-core
+
+## igniteui/angular-templates@13.2.920
+
+* Update to igniteui-angular 13.2.0 by @dafo in https://github.com/IgniteUI/igniteui-cli/pull/1026
+
+## igniteui/angular-schematics@13.2.920
+
+* Note: Version bump only for package igniteui/angular-schematics
+
+## igniteui-cli@9.2.0
+
+* Update to igniteui-angular 13.2.0 by @dafo in https://github.com/IgniteUI/igniteui-cli/pull/1026
+
+# [9.1.4](https://github.com/IgniteUI/igniteui-cli/compare/v9.1.3...v9.1.4)  (2022-05-20)
+
+## igniteui/cli-core@9.1.4
+
+* Note: Version bump only for package igniteui/cli-core
+
+## igniteui/angular-templates@13.1.914
+
+* Note: Version bump only for package igniteui/angular-templates
+
+## igniteui/angular-schematics@13.1.914
+
+#### Bug Fixes
+
+* **schematics:** refactor(theme-import): Place sass/scss imports before content.
+ ([fb5e4d3](https://github.com/IgniteUI/igniteui-cli/commit/fb5e4d3c217c68f815afa753120fa3673cf79a06))
+
+## igniteui-cli@9.1.4
+
+#### Bug Fixes
+
+* **schematics:** refactor(theme-import): Place sass/scss imports before content.
+ ([fb5e4d3](https://github.com/IgniteUI/igniteui-cli/commit/fb5e4d3c217c68f815afa753120fa3673cf79a06))
+
+# [9.1.3](https://github.com/IgniteUI/igniteui-cli/compare/v9.1.2...v9.1.3)  (2022-05-19)
+
+## igniteui/cli-core@9.1.3
+
+* Note: Version bump only for package igniteui/cli-core
+
+## igniteui/angular-templates@13.1.913
+
+* Note: Version bump only for package igniteui/angular-templates
+
+## igniteui/angular-schematics@13.1.913
+
+#### Bug Fixes
+
+* **schematics:** Use @use rule to add default theming module [#1023](https://github.com/IgniteUI/igniteui-cli/issues/1023) ([fc16535](https://github.com/IgniteUI/igniteui-cli/commit/fc165358d0aa368e592195ae0666511d095b5d0c))
+
+## igniteui-cli@9.1.3
+
+#### Bug Fixes
+
+* **schematics:** Use @use rule to add default theming module [#1023](https://github.com/IgniteUI/igniteui-cli/issues/1023) ([fc16535](https://github.com/IgniteUI/igniteui-cli/commit/fc165358d0aa368e592195ae0666511d095b5d0c))
+
+# [9.1.2](https://github.com/IgniteUI/igniteui-cli/compare/v9.1.1...v9.1.2)  (2022-05-16)
+
+## igniteui/cli-core@9.1.2
+
+* Note: Version bump only for package igniteui/cli-core
+
+## igniteui/angular-templates@13.1.912
+
+#### Bug Fixes
+
+* **igx(fintech grid):** fix slow grid [#1020](https://github.com/IgniteUI/igniteui-cli/issues/1020) ([4f888dc](https://github.com/IgniteUI/igniteui-cli/commit/4f888dc3c2b1fd6ad896b197634d3aa0c3ef57cd))
+
+## igniteui/angular-schematics@13.1.912
+
+* Note: Version bump only for package igniteui/angular-schematics
+
+## igniteui-cli@9.1.2
+
+* **igx(fintech grid):** fix slow grid [#1020](https://github.com/IgniteUI/igniteui-cli/issues/1020) ([4f888dc](https://github.com/IgniteUI/igniteui-cli/commit/4f888dc3c2b1fd6ad896b197634d3aa0c3ef57cd))
+
+# [9.1.1](https://github.com/IgniteUI/igniteui-cli/compare/v9.1.0...v9.1.1)  (2022-03-18)
+
+🎉 This update includes:
+
+- `igniteui-webcomponents` version [2.1.0](https://github.com/IgniteUI/igniteui-webcomponents/releases/tag/2.1.0).
+
+## igniteui/cli-core@9.1.1
+
+* Note: Version bump only for package igniteui/cli-core
+
+## igniteui/angular-templates@13.1.911
+
+* update with @angular@13.3 by @Lipata in https://github.com/IgniteUI/igniteui-cli/pull/1015
+* Template for the pivot grid by @valadzhov in https://github.com/IgniteUI/igniteui-cli/pull/1013
+
+## igniteui/angular-schematics@13.1.911
+
+* Note: Version bump only for package igniteui/angular-schematics
+
+## igniteui-cli@9.1.1
+
+* feat(wc): add new 2.1.0 components by @Lipata in https://github.com/IgniteUI/igniteui-cli/pull/1014
+* update with @angular@13.3 by @Lipata in https://github.com/IgniteUI/igniteui-cli/pull/1015
+* Template for the pivot grid by @valadzhov in https://github.com/IgniteUI/igniteui-cli/pull/1013
+
+# [9.1.0](https://github.com/IgniteUI/igniteui-cli/compare/v9.0.6...v9.1.0)  (2022-03-02)
+
+🎉 This update includes:
+
+- `igniteui-angular` version [13.1.0](https://github.com/IgniteUI/igniteui-angular/releases/tag/13.1.0).
+
+## igniteui/cli-core@9.1.0
+
+* Note: Version bump only for package igniteui/cli-core
+
+## igniteui/angular-templates@13.1.910
+
+* Update the templates in accordance with the latest migrations for 13.1 by @valadzhov in https://github.com/IgniteUI/igniteui-cli/pull/1006
+
+## igniteui/angular-schematics@13.1.910
+
+* Note: Version bump only for package igniteui/angular-schematics
+
+## igniteui-cli@9.1.0
+
+* Adding Empty view for WC project by @dafo in https://github.com/IgniteUI/igniteui-cli/pull/979
+* Using rollup as building tool in Web components projects by @dafo in https://github.com/IgniteUI/igniteui-cli/pull/999
+* Pass in options for wc child process by @jackofdiamond5 in https://github.com/IgniteUI/igniteui-cli/pull/1000
+* Web Components templates use LitElement by @dafo in https://github.com/IgniteUI/igniteui-cli/pull/1001
+* Update the templates in accordance with the latest migrations for 13.1 by @valadzhov in https://github.com/IgniteUI/igniteui-cli/pull/1006
+
 # [9.0.6](https://github.com/IgniteUI/igniteui-cli/compare/v9.0.5...v9.0.6)  (2022-02-04)
 
 ## igniteui/cli-core@9.0.6